integration with inventory service using feign client

+++ src/main/java/com/example/orders/core/application/dto/ProductDetailsDto.java
@@ -5,14 +5,14 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 @JsonInclude(JsonInclude.Include.NON_NULL)  // Excludes null values
 public class ProductDetailsDto {
     private String productName;
-    private Long productId;
+    private Integer productId;
     private Integer quantity;
     private Double priceOfUnit;
 
     public ProductDetailsDto() {
     }
 
-    public ProductDetailsDto(String productName, Long productId, Integer quantity, Double priceOfUnit) {
+    public ProductDetailsDto(String productName, Integer productId, Integer quantity, Double priceOfUnit) {
         this.productName = productName;
         this.productId = productId;
         this.quantity = quantity;
@@ -27,11 +27,11 @@ public class ProductDetailsDto {
         this.productName = productName;
     }
 
-    public Long getProductId() {
+    public Integer getProductId() {
         return productId;
     }
 
-    public void setProductId(Long productId) {
+    public void setProductId(Integer productId) {
         this.productId = productId;
     }
